+++ backend/app/income/models.py
@@ -3,7 +3,9 @@ import enum
 import uuid
 from datetime import date
 
-from sqlalchemy import Date, Enum, Float, ForeignKey, String, Text
+from decimal import Decimal
+
+from sqlalchemy import Date, Enum, ForeignKey, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base, TimestampMixin
@@ -35,7 +37,7 @@ class Income(TimestampMixin, Base):
         Enum(IncomeCategory), default=IncomeCategory.OTHER, nullable=False
     )
     description: Mapped[str] = mapped_column(String(1000), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="USD", nullable=False)
     date: Mapped[date] = mapped_column(Date, nullable=False, index=True)
     payment_method: Mapped[str | None] = mapped_column(String(50), nullable=True)
